+++ src/test/java/info/avalon566/shardingscaling/sync/mysql/binlog/codec/DataTypesCodecTest.java
@@ -84,15 +84,15 @@ public class DataTypesCodecTest {
     @Test
     public void assertReadInt1() {
         when(byteBuf.readByte()).thenReturn(Byte.MAX_VALUE, Byte.MIN_VALUE);
-        assertThat(DataTypesCodec.readInt1(byteBuf), is((int) Byte.MAX_VALUE));
-        assertThat(DataTypesCodec.readInt1(byteBuf), is((int) Byte.MIN_VALUE));
+        assertThat(DataTypesCodec.readInt1(byteBuf), is(Byte.MAX_VALUE));
+        assertThat(DataTypesCodec.readInt1(byteBuf), is(Byte.MIN_VALUE));
     }
 
     @Test
     public void assertReadInt2LE() {
         when(byteBuf.readShortLE()).thenReturn(Short.MAX_VALUE, Short.MIN_VALUE);
-        assertThat(DataTypesCodec.readInt2LE(byteBuf), is((int) Short.MAX_VALUE));
-        assertThat(DataTypesCodec.readInt2LE(byteBuf), is((int) Short.MIN_VALUE));
+        assertThat(DataTypesCodec.readInt2LE(byteBuf), is(Short.MAX_VALUE));
+        assertThat(DataTypesCodec.readInt2LE(byteBuf), is(Short.MIN_VALUE));
     }
 
     @Test
